Feature legg til beregningslogikk

+++ src/main/java/no/nav/bidrag/beregn/bidragsevne/beregning/BidragsevneberegningImpl.java
@@ -122,7 +122,7 @@ public class BidragsevneberegningImpl implements Bidragsevneberegning {
     // Trekker fra midler til underhold egne barn i egen husstand
     forelopigBidragsevne = forelopigBidragsevne.subtract(
         sjablonNavnVerdiMap.get(SjablonTallNavn.UNDERHOLD_EGNE_BARN_I_HUSSTAND_BELOP.getNavn())
-            .multiply(BigDecimal.valueOf(grunnlagBeregning.getAntallEgneBarnIHusstand()))
+            .multiply(grunnlagBeregning.getAntallEgneBarnIHusstand())
             .multiply(BigDecimal.valueOf(12)));
 //    System.out.println("Foreløpig evne etter fratrekk av underhold for egne barn i egen husstand: "
 //        + forelopigBidragsevne);
