+++ src/test/java/no/nav/bidrag/beregn/bidragsevne/BidragsevneberegningTest.java
@@ -38,14 +38,14 @@ class BidragsevneberegningTest {
     // Tester beregning med ulike inntekter
     inntekter.add(new Inntekt(InntektType.LONN_SKE, BigDecimal.valueOf(1000000)));
     GrunnlagBeregning grunnlagBeregning
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 1,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(1),
         SaerfradragKode.INGEN, sjablonListe);
     assertEquals(BigDecimal.valueOf(31859),
         bidragsevneberegning.beregn(grunnlagBeregning).getResultatEvneBelop());
 
     inntekter.set(0, new Inntekt(InntektType.LONN_SKE, BigDecimal.valueOf(520000)));
     GrunnlagBeregning grunnlagBeregning2
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 1,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(1),
         SaerfradragKode.INGEN, sjablonListe);
     assertEquals(BigDecimal.valueOf(8322),
         bidragsevneberegning.beregn(grunnlagBeregning2).getResultatEvneBelop());
@@ -54,7 +54,7 @@ class BidragsevneberegningTest {
 
     inntekter.set(0, new Inntekt(InntektType.LONN_SKE, BigDecimal.valueOf(666000)));
     GrunnlagBeregning grunnlagBeregning3
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 3,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(3),
         SaerfradragKode.INGEN, sjablonListe);
     assertEquals(BigDecimal.valueOf(8424),
         bidragsevneberegning.beregn(grunnlagBeregning3).getResultatEvneBelop());
@@ -62,7 +62,7 @@ class BidragsevneberegningTest {
     // Test på at beregnet bidragsevne blir satt til 0 når evne er negativ
     inntekter.set(0, new Inntekt(InntektType.LONN_SKE, BigDecimal.valueOf(100000)));
     GrunnlagBeregning grunnlagBeregning4
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.MED_ANDRE, 1,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.MED_ANDRE, BigDecimal.valueOf(1),
         SaerfradragKode.HELT, sjablonListe);
     assertEquals(BigDecimal.ZERO,
         bidragsevneberegning.beregn(grunnlagBeregning4).getResultatEvneBelop());
@@ -73,7 +73,7 @@ class BidragsevneberegningTest {
         singletonList(new SjablonInnhold(SjablonInnholdNavn.SJABLON_VERDI.getNavn(), BigDecimal.valueOf(12000)))));
 
     GrunnlagBeregning grunnlagBeregning5
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 3,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(3),
         SaerfradragKode.INGEN, sjablonListe);
     assertEquals(BigDecimal.valueOf(8424),
         bidragsevneberegning.beregn(grunnlagBeregning5).getResultatEvneBelop());
@@ -84,7 +84,7 @@ class BidragsevneberegningTest {
         singletonList(new SjablonInnhold(SjablonInnholdNavn.SJABLON_VERDI.getNavn(), BigDecimal.valueOf(12000)))));
 
     GrunnlagBeregning grunnlagBeregning6
-        = new GrunnlagBeregning(inntekter, 2, BostatusKode.ALENE, 3,
+        = new GrunnlagBeregning(inntekter, 2, BostatusKode.ALENE, BigDecimal.valueOf(3),
         SaerfradragKode.INGEN, sjablonListe);
     assertEquals(BigDecimal.valueOf(9424),
         bidragsevneberegning.beregn(grunnlagBeregning6).getResultatEvneBelop());
@@ -97,7 +97,7 @@ class BidragsevneberegningTest {
         singletonList(new SjablonInnhold(SjablonInnholdNavn.SJABLON_VERDI.getNavn(), BigDecimal.valueOf(24000)))));
 
     GrunnlagBeregning grunnlagBeregning7
-        = new GrunnlagBeregning(inntekter, 2, BostatusKode.ALENE, 3,
+        = new GrunnlagBeregning(inntekter, 2, BostatusKode.ALENE, BigDecimal.valueOf(3),
         SaerfradragKode.INGEN, sjablonListe);
     assertEquals(BigDecimal.valueOf(7923),
         bidragsevneberegning.beregn(grunnlagBeregning7).getResultatEvneBelop());
@@ -105,14 +105,14 @@ class BidragsevneberegningTest {
 
     // Test av halvt særfradrag
     GrunnlagBeregning grunnlagBeregning8
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 3,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(3),
         SaerfradragKode.HALVT, sjablonListe);
     assertEquals(BigDecimal.valueOf(8965),
         bidragsevneberegning.beregn(grunnlagBeregning8).getResultatEvneBelop());
 
     // Test av bostatus MED_FLERE
     GrunnlagBeregning grunnlagBeregning9
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.MED_ANDRE, 3,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.MED_ANDRE, BigDecimal.valueOf(3),
         SaerfradragKode.HALVT, sjablonListe);
     assertEquals(BigDecimal.valueOf(14253),
         bidragsevneberegning.beregn(grunnlagBeregning9).getResultatEvneBelop());
@@ -128,7 +128,7 @@ class BidragsevneberegningTest {
     BidragsevneberegningImpl bidragsevneberegning = new BidragsevneberegningImpl();
 
     GrunnlagBeregning grunnlagBeregning
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 1,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(1),
         SaerfradragKode.HELT, sjablonListe);
     System.out.println(bidragsevneberegning.beregnMinstefradrag(grunnlagBeregning,
         SjablonUtil.hentSjablonverdi(sjablonListe, SjablonTallNavn.MINSTEFRADRAG_INNTEKT_BELOP),
@@ -140,7 +140,7 @@ class BidragsevneberegningTest {
 
     inntekter.set(0, new Inntekt(InntektType.LONN_SKE, BigDecimal.valueOf(1000000)));
     GrunnlagBeregning grunnlagBeregning2
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 1,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(1),
         SaerfradragKode.HELT, sjablonListe);
     System.out.println(bidragsevneberegning.beregnMinstefradrag(grunnlagBeregning2,
         SjablonUtil.hentSjablonverdi(sjablonListe, SjablonTallNavn.MINSTEFRADRAG_INNTEKT_BELOP),
@@ -160,21 +160,21 @@ class BidragsevneberegningTest {
     inntekter.add(new Inntekt(InntektType.LONN_SKE, BigDecimal.valueOf(666000)));
 
     GrunnlagBeregning grunnlagBeregning
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 1,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(1),
         SaerfradragKode.HELT, sjablonListe);
     assertEquals(BigDecimal.valueOf(1400+16181+3465+0),
         bidragsevneberegning.beregnSkattetrinnBelop(grunnlagBeregning));
 
     inntekter.set(0, new Inntekt(InntektType.LONN_SKE, BigDecimal.valueOf(174600)));
     GrunnlagBeregning grunnlagBeregning2
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 1,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(1),
         SaerfradragKode.HELT, sjablonListe);
     assertEquals(BigDecimal.ZERO,
         bidragsevneberegning.beregnSkattetrinnBelop(grunnlagBeregning2));
 
     inntekter.set(0, new Inntekt(InntektType.LONN_SKE, BigDecimal.valueOf(250000)));
     GrunnlagBeregning grunnlagBeregning3
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 1,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(1),
         SaerfradragKode.HELT, sjablonListe);
     assertEquals(BigDecimal.valueOf(1315),
         bidragsevneberegning.beregnSkattetrinnBelop(grunnlagBeregning3));
@@ -190,7 +190,7 @@ class BidragsevneberegningTest {
     // Tester beregning med ulike inntekter
     inntekter.add(new Inntekt(InntektType.LONN_SKE, BigDecimal.valueOf(300000)));
     GrunnlagBeregning grunnlagBeregning
-        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, 0,
+        = new GrunnlagBeregning(inntekter, 1, BostatusKode.ALENE, BigDecimal.valueOf(0),
         SaerfradragKode.HELT, sjablonListe);
     assertEquals(BigDecimal.valueOf(1217),
         bidragsevneberegning.beregn(grunnlagBeregning).getResultatEvneBelop());
